car_rentals_backend/requirements.txt
fastapi[all]
uvicorn
pymongo
jose
passlib
python-jose[cryptography]
pydantic-settings
bcrypt